+++ packages/testing-tachyons-rn/src/Css.test.tsx
@@ -1,0 +1,11 @@
+import { Css } from "./Css";
+
+describe("Css", () => {
+  it("can add mb", () => {
+    expect(Css.mb1.$).toMatchInlineSnapshot(`
+      {
+        "marginBottom": "8px",
+      }
+    `);
+  });
+});