+++ packages/truss/src/sections/tachyons/index.ts
@@ -31,6 +31,8 @@ import { whitespace } from "src/sections/tachyons/whitespace";
 import { width } from "src/sections/tachyons/widths";
 import { wordBreak } from "src/sections/tachyons/wordBreak";
 import { zIndex } from "src/sections/tachyons/zIndex";
+import { container } from "src/sections/tachyons/container";
+import { opacity } from "src/sections/tachyons/opacity";
 
 export const defaultSections = {
   border,
@@ -39,6 +41,7 @@ export const defaultSections = {
   borderStyle,
   borderWidth,
   boxShadow,
+  container,
   coordinates,
   cursor,
   display,
@@ -49,6 +52,7 @@ export const defaultSections = {
   height,
   lineClamp,
   objectFit,
+  opacity,
   outline,
   overflow,
   position,