+++ packages/truss/src/sections/tachyons/textDecoration.ts
@@ -3,7 +3,7 @@ import { CreateMethodsFn } from "src/config";
 
 export const textDecoration: CreateMethodsFn = () =>
   newMethodsForProp("textDecoration", {
-    noUnderline: "none",
-    strike: "line-through",
-    underline: "underline",
+    tdn: "none",
+    tdlt: "line-through",
+    tdu: "underline",
   });