+++ packages/truss/src/sections/tachyons/typeScale.ts
@@ -1,10 +1,12 @@
-import { newMethod } from "src/methods";
+import { newMethod, newMethodsForProp, newParamMethod } from "src/methods";
 import { CreateMethodsFn } from "src/config";
 
-export const typeScale: CreateMethodsFn = ({ fonts }) =>
-  Object.entries(fonts).map(([abbr, defs]) => {
+/** Adds abbreviations like `tiny` -> `12px` or `tiny` -> `{ fontSize: 12px, fontWeight: ... }`. */
+export const typeScale: CreateMethodsFn = ({ fonts }) => [
+  ...Object.entries(fonts).map(([abbr, defs]) => {
     if (typeof defs === "string") {
       return newMethod(abbr, { fontSize: defs });
     }
     return newMethod(abbr, defs);
-  });
+  }),
+];