+++ CHANGELOG.md
@@ -1,3 +1,185 @@
+# [1.136.0](https://github.com/homebound-team/truss/compare/v1.135.0...v1.136.0) (2024-06-18)
+
+
+### Features
+
+* Add fs/fsPx for fontSize. ([#122](https://github.com/homebound-team/truss/issues/122)) ([a4b541e](https://github.com/homebound-team/truss/commit/a4b541e16ff3cc8627c7ee091dfbddc16b7ff4d4))
+
+# [1.135.0](https://github.com/homebound-team/truss/compare/v1.134.0...v1.135.0) (2024-05-31)
+
+
+### Features
+
+* Double down on abbreviation style. ([#121](https://github.com/homebound-team/truss/issues/121)) ([6b28580](https://github.com/homebound-team/truss/commit/6b28580ee0aae3afb4b443e20650236e33307abd))
+
+# [1.134.0](https://github.com/homebound-team/truss/compare/v1.133.0...v1.134.0) (2024-05-20)
+
+
+### Features
+
+* support additional properties to utility method ([#120](https://github.com/homebound-team/truss/issues/120)) ([a92172d](https://github.com/homebound-team/truss/commit/a92172d6ea16a2c6be30d855b8e64592e6381a76))
+
+# [1.133.0](https://github.com/homebound-team/truss/compare/v1.132.0...v1.133.0) (2024-03-05)
+
+
+### Features
+
+* Add order and opacity shorthand styles ([#119](https://github.com/homebound-team/truss/issues/119)) ([c5e5d2c](https://github.com/homebound-team/truss/commit/c5e5d2cf25eb7229feb0e05d1926c33ab4802f90))
+
+# [1.132.0](https://github.com/homebound-team/truss/compare/v1.131.1...v1.132.0) (2023-12-06)
+
+
+### Features
+
+* Add flex-wrap styles ([#118](https://github.com/homebound-team/truss/issues/118)) ([698070c](https://github.com/homebound-team/truss/commit/698070c3a77a38a7720f4f3604bc2f51978e4fb9))
+
+## [1.131.1](https://github.com/homebound-team/truss/compare/v1.131.0...v1.131.1) (2023-11-08)
+
+
+### Bug Fixes
+
+* CircleCI yarn cache [sc-42645] ([#117](https://github.com/homebound-team/truss/issues/117)) ([975f68f](https://github.com/homebound-team/truss/commit/975f68fa402de5bc1d2b9fcfac24e101dc74e269))
+
+# [1.131.0](https://github.com/homebound-team/truss/compare/v1.130.0...v1.131.0) (2023-08-16)
+
+
+### Features
+
+* Add mwfc. ([#114](https://github.com/homebound-team/truss/issues/114)) ([49c90de](https://github.com/homebound-team/truss/commit/49c90de23267e7cf12565727dbd1a2e5015ee4e7))
+
+# [1.130.0](https://github.com/homebound-team/truss/compare/v1.129.0...v1.130.0) (2023-08-11)
+
+
+### Features
+
+* Add tachyons-rn placeholder. ([#113](https://github.com/homebound-team/truss/issues/113)) ([67cc458](https://github.com/homebound-team/truss/commit/67cc4586da06e06f1e2bb667de02ce3ac7c04b0f))
+
+# [1.129.0](https://github.com/homebound-team/truss/compare/v1.128.1...v1.129.0) (2023-07-18)
+
+
+### Features
+
+* Add support for container queries ([#111](https://github.com/homebound-team/truss/issues/111)) ([2218354](https://github.com/homebound-team/truss/commit/221835425eedc2b2348d7fea58b20c6e40fe46cd))
+
+## [1.128.1](https://github.com/homebound-team/truss/compare/v1.128.0...v1.128.1) (2023-07-01)
+
+
+### Bug Fixes
+
+* Export newPxMethod. Fixes [#100](https://github.com/homebound-team/truss/issues/100). ([#108](https://github.com/homebound-team/truss/issues/108)) ([4dc329c](https://github.com/homebound-team/truss/commit/4dc329c1746017ba650a52a659736d5f40091028))
+
+# [1.128.0](https://github.com/homebound-team/truss/compare/v1.127.0...v1.128.0) (2023-07-01)
+
+
+### Features
+
+* Add placeItems abbreviations. ([#107](https://github.com/homebound-team/truss/issues/107)) ([9ab948b](https://github.com/homebound-team/truss/commit/9ab948b2347618be9c7977de2ce210490336b10a))
+
+# [1.127.0](https://github.com/homebound-team/truss/compare/v1.126.1...v1.127.0) (2023-05-08)
+
+
+### Features
+
+* Better support for data attributes ([#106](https://github.com/homebound-team/truss/issues/106)) ([bf7c0b2](https://github.com/homebound-team/truss/commit/bf7c0b25575034145a261f897bfcc647135670ce))
+
+## [1.126.1](https://github.com/homebound-team/truss/compare/v1.126.0...v1.126.1) (2023-03-27)
+
+
+### Bug Fixes
+
+* Actually codegen. ([1704cfa](https://github.com/homebound-team/truss/commit/1704cfa6943f87b0cf687f050a25d1e0d935d0bf))
+
+# [1.126.0](https://github.com/homebound-team/truss/compare/v1.125.0...v1.126.0) (2023-03-27)
+
+
+### Features
+
+* Add whiteSpace: breakspaces. ([#104](https://github.com/homebound-team/truss/issues/104)) ([793251d](https://github.com/homebound-team/truss/commit/793251dea330763f03268fb580b0116f5c8b8570))
+
+# [1.125.0](https://github.com/homebound-team/truss/compare/v1.124.0...v1.125.0) (2023-03-08)
+
+
+### Features
+
+* Support grid-auto-columns and grid-auto-rows ([#103](https://github.com/homebound-team/truss/issues/103)) ([1519d4c](https://github.com/homebound-team/truss/commit/1519d4c6152532a6d5dce543da60b5b4f37e49e4))
+
+# [1.124.0](https://github.com/homebound-team/truss/compare/v1.123.0...v1.124.0) (2023-03-04)
+
+
+### Features
+
+* Add width/height max/min/fit-content abbreviations. ([#102](https://github.com/homebound-team/truss/issues/102)) ([c99a56b](https://github.com/homebound-team/truss/commit/c99a56b494b1ffcc6f5b3476f7926ccb600336a7))
+
+# [1.123.0](https://github.com/homebound-team/truss/compare/v1.122.1...v1.123.0) (2023-03-03)
+
+
+### Features
+
+* Add place-self. ([#101](https://github.com/homebound-team/truss/issues/101)) ([19081a9](https://github.com/homebound-team/truss/commit/19081a9ae89e3e68dc11ab7a3e27811804f96d6a))
+
+## [1.122.1](https://github.com/homebound-team/truss/compare/v1.122.0...v1.122.1) (2023-01-11)
+
+
+### Bug Fixes
+
+* Add wa/ha for width/height auto. ([#95](https://github.com/homebound-team/truss/issues/95)) ([dc10094](https://github.com/homebound-team/truss/commit/dc10094543a8904ec60579e772b9157889ec9ed4))
+
+# [1.122.0](https://github.com/homebound-team/truss/compare/v1.121.1...v1.122.0) (2023-01-11)
+
+
+### Features
+
+* Add 'auto' margin abbreviations. ([#94](https://github.com/homebound-team/truss/issues/94)) ([0a6598b](https://github.com/homebound-team/truss/commit/0a6598b003a7f6b5340cd53514765b588241de4b))
+
+## [1.121.1](https://github.com/homebound-team/truss/compare/v1.121.0...v1.121.1) (2022-11-28)
+
+
+### Bug Fixes
+
+* Restore WebkitBoxOrient. ([#91](https://github.com/homebound-team/truss/issues/91)) ([5729145](https://github.com/homebound-team/truss/commit/5729145aee6f31b3eb5260edc52930a562df0a37))
+
+# [1.121.0](https://github.com/homebound-team/truss/compare/v1.120.0...v1.121.0) (2022-11-28)
+
+
+### Features
+
+* Support else w/breakpoints, fix addIn w/conditionals ([#90](https://github.com/homebound-team/truss/issues/90)) ([eecacee](https://github.com/homebound-team/truss/commit/eecacee507ae18eff56fccd3eaf62503207c2703))
+
+# [1.120.0](https://github.com/homebound-team/truss/compare/v1.119.0...v1.120.0) (2022-11-23)
+
+
+### Features
+
+* Add comments to the utility methods. ([#89](https://github.com/homebound-team/truss/issues/89)) ([d0f6720](https://github.com/homebound-team/truss/commit/d0f67209eb7f759c6b39d72df747513af3b0d265))
+
+# [1.119.0](https://github.com/homebound-team/truss/compare/v1.118.0...v1.119.0) (2022-11-23)
+
+
+### Features
+
+* Simplify the exported Properties type. ([#88](https://github.com/homebound-team/truss/issues/88)) ([4570ff3](https://github.com/homebound-team/truss/commit/4570ff333635c5380a236343d9bfee0cc55f40ac))
+
+# [1.118.0](https://github.com/homebound-team/truss/compare/v1.117.0...v1.118.0) (2022-11-23)
+
+
+### Features
+
+* Add sqPx for setting w and h. ([#87](https://github.com/homebound-team/truss/issues/87)) ([74b3f1d](https://github.com/homebound-team/truss/commit/74b3f1dc459c2e3b7a0e43f42758779256d139e8))
+
+# [1.117.0](https://github.com/homebound-team/truss/compare/v1.116.0...v1.117.0) (2022-11-23)
+
+
+### Features
+
+* Add onHover. ([#86](https://github.com/homebound-team/truss/issues/86)) ([fc8e73a](https://github.com/homebound-team/truss/commit/fc8e73a88a048999e0303de7a491775e6945b9e6))
+
+# [1.116.0](https://github.com/homebound-team/truss/compare/v1.115.0...v1.116.0) (2022-11-23)
+
+
+### Features
+
+* Add if<breakpoint> methods. ([#85](https://github.com/homebound-team/truss/issues/85)) ([591cc3e](https://github.com/homebound-team/truss/commit/591cc3e524018310a235d9d33d9a23892edfb802))
+
 # [1.115.0](https://github.com/homebound-team/truss/compare/v1.114.0...v1.115.0) (2022-10-29)